% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom_line_indices.R
\name{bottom_line_indices}
\alias{bottom_line_indices}
\title{Get bottom line indices of a given plateformat}
\usage{
bottom_line_indices(plateformat)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536}
}
\value{
The numeric vector with bottom line indices of  a given
\code{plateformat}
}
\description{
Get bottom line indices of a given plateformat
}
\examples{
bottom_line_indices(384)
bottom_line_indices(1536)
}

