% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom_line_neighbors.R
\name{bottom_line_neighbors}
\alias{bottom_line_neighbors}
\title{Get indices of five neighbors for an index at the bottom line of the plate}
\usage{
bottom_line_neighbors(plateformat, n)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536}

\item{n}{An integer which can be one of \code{bottom_line_indices(plateformat)}}
}
\value{
numeric vector of five neighbors for the given \code{plateformat} and
\code{n}
}
\description{
Get indices of five neighbors for an index at the bottom line of the plate
}
\examples{
bottom_line_neighbors(384, 32)
bottom_line_neighbors(1536, 1504)
}

