% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_data.R
\name{categorize_data}
\alias{categorize_data}
\title{Categorize data based on plate median}
\usage{
categorize_data(data_area, empty_indices)
}
\arguments{
\item{data_area}{A numerical vector giving the colony size of a
plate}

\item{empty_indices}{A numerical vector giving the indices of
empty spot locations on a plate}
}
\value{
A character vector with different classes of data
}
\description{
The function \code{categorize_data} will classify the given data
\code{data_area} on the basis of plate median and after removing
missing values \code{NA}.
}
\examples{
data_area <- simulated_data_1536(data_384 = colonyarea$data_subtypes,
                                 in_data_flow = "across",
                                 out_data_flow = "down",
                                 is_plate_coords = FALSE)

empty_indices <- which(convert_small_to_large(plate_from = 384,
                                              plate_to = 1536,
                                              data_from = colonyarea$data_subtypes,
                                              in_data_flow = 'across',
                                              out_data_flow = "down",
                                              is_plate_coords = FALSE)$y \%in\% 'Empty')

vec1 <- categorize_data(data_area$y, empty_indices)
head(vec1)

}

