% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colonyarea}
\alias{colonyarea}
\title{Yeast Colony Area on 384 Plate Format}
\format{A data frame with 384 rows and 3 variables.

\tabular{rll}{
 [, 1] \tab \code{row}           \tab Horizontal coordinates of a plate.
                                      It must have character class.\cr
 [, 2] \tab \code{column}        \tab Vertical coordinates of a plate.
                                      It must have numeric class.\cr
 [, 3] \tab \code{data_subtypes} \tab It must have character class
                                      indicating the data subtypes.\cr
}}
\usage{
colonyarea
}
\description{
A dataset containing the plate coordinates, and data subtypes obtained by
categorizing the colony area of yeast using plate growth median
on a 384 plate format.
}
\author{
Sathish Kumar Srinivasan  \email{sathishsrinivasank@gmail.com}
}
\keyword{datasets}

