% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_down_across.R
\name{convert_down_across}
\alias{convert_down_across}
\title{Convert data into rowwise(across) or coloumnwise(down) within a plate format}
\usage{
convert_down_across(plateformat, data_from, is_plate_coords = TRUE,
  out_data_flow)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536 or 6144}

\item{data_from}{vector of data for conversion}

\item{is_plate_coords}{logical returns plate coordinates, if its value is
\code{TRUE}.}

\item{out_data_flow}{A string which tells the function to output the
converted data to a plate configuration which can be down or across.}
}
\value{
dataframe of converted data from across to down or vice versa.
}
\description{
Convert data into rowwise(across) or coloumnwise(down) within a plate format
}
\details{
Make sure the length of the given data in \code{data_from} argument matches
the value in \code{plateformat}
}
\examples{
plateformat <- 384
convert_down_across(plateformat = plateformat,
                    data_from = 1:plateformat,
                    is_plate_coords = TRUE,
                    out_data_flow = 'across')
}

