% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_large_to_small.R
\name{convert_large_to_small}
\alias{convert_large_to_small}
\title{Convert data of large plateformat to small plateformat}
\usage{
convert_large_to_small(plate_from, plate_to, data_from, in_data_flow,
  out_data_flow, is_plate_coords)
}
\arguments{
\item{plate_from}{An integer of large plateformat}

\item{plate_to}{An integer of small plateformat}

\item{data_from}{vector of data for conversion}

\item{in_data_flow}{A string indicating the plate format of \code{data_from}.
It can take a value either across or down.}

\item{out_data_flow}{A string which tells the function to output the
converted data to a plate configuration which can be down or across}

\item{is_plate_coords}{logical returns plate coordinates, if its value is
\code{TRUE}}
}
\value{
dataframe of converted data from large plateformat \code{plate_from}
to small plateformat \code{plate_to}. If \code{is_plate_coords = TRUE}, then
it returns the plate coordinates along with the converted data
}
\description{
Convert data of large plateformat  \code{plate_from} to small plateformat
\code{plate_to} by getting the four replicates of large plateformat
\code{plate_from}
}
\details{
Always convert the plateformat one step at a time. For example 1536 to 384
but not from 1536 to 96 plateformat
}
\examples{
down_384 <- convert_large_to_small(plate_from = 1536,
                                   plate_to = 384,
                                   data_from = 1:1536,
                                   in_data_flow = 'down',
                                   out_data_flow = 'down',
                                   is_plate_coords = TRUE)
head(down_384)

}

