% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_small_to_large.R
\name{convert_small_to_large}
\alias{convert_small_to_large}
\title{Convert data of small plateformat to large plateformat}
\usage{
convert_small_to_large(plate_from, plate_to, data_from, out_data_flow,
  is_plate_coords)
}
\arguments{
\item{plate_from}{An integer of small plateformat}

\item{plate_to}{An integer of large plateformat}

\item{data_from}{vector of data for conversion}

\item{out_data_flow}{A string which tells the function to output the
converted data to a plate configuration which can be down or across}

\item{is_plate_coords}{logical returns plate coordinates, if its value is
\code{TRUE}}
}
\value{
dataframe of converted data from small plateformat \code{plate_from}
to large plateformat \code{plate_to}. If \code{is_plate_coords = TRUE}, then
it returns the plate coordinates along with the converted data.
}
\description{
Convert data of small plateformat \code{plate_from} to large plateformat
\code{plate_to} by expanding the four replicates into large plateformat
\code{plate_to}
}
\examples{
across_1536 <- convert_small_to_large(plate_from = 384,
                                      plate_to = 1536,
                                      data_from = colonyarea$data_subtypes,
                                      out_data_flow = 'across',
                                      is_plate_coords = TRUE)
head(across_1536)

}

