% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_4_replicates.R
\name{indices_4_replicates}
\alias{indices_4_replicates}
\title{Indices of four replicates from large to small plateformat having across or
down configuration}
\usage{
indices_4_replicates(plate_from, plate_to, out_data_flow)
}
\arguments{
\item{plate_from}{An integer of large plateformat}

\item{plate_to}{An integer of small plateformat}

\item{out_data_flow}{A string which tells the function to output the
converted data to a plate configuration which can be down or across}
}
\value{
dataframe with indices of four replicates converted from large to
small plateformat with the configuration specified in \code{out_data_flow}
}
\description{
Indices of four replicates from large to small plateformat having across or
down configuration
}
\examples{
indices_384_96 <- indices_4_replicates(plate_from    = 384,
                                       plate_to      = 96,
                                       out_data_flow = 'across')
head(indices_384_96)
}

