% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_line_indices.R
\name{left_line_indices}
\alias{left_line_indices}
\title{Get left line indices of a given plateformat}
\usage{
left_line_indices(plateformat)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536}
}
\value{
The numeric vector with left line indices of  a given
\code{plateformat}
}
\description{
Get left line indices of a given plateformat
}
\examples{
left_line_indices(384)
left_line_indices(1536)
}

