% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborless_algorithm.R
\name{neighborless_algorithm}
\alias{neighborless_algorithm}
\title{Neighborless Algorithm}
\usage{
neighborless_algorithm(plateformat, colony_area_raw_data, empty_indices,
  excluded_colonies, is_save = FALSE, excluded_file = NULL)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536.}

\item{colony_area_raw_data}{A numeric vector of raw data representing the area
yeast grown at a specific location on a nutrient medium agar plate. The data can
be either rowwise or columnwise and its arrangement does not influence the
outcome of this algorithm.}

\item{empty_indices}{A numeric vector of indexed empty locations. It can also
be the indices of control colonies grown on a plate, which may be used for
normalizing the colony area of a plate}

\item{excluded_colonies}{A numeric vector of indices that are part of excluded
colonies}

\item{is_save}{logical. If \code{TRUE}, then this function will save the
list of excluded colonies obtained from neighborless algorithm as .RDS format
in the specified \code{excluded_file}.}

\item{excluded_file}{If \code{is_save} is \code{TRUE}, then the file name
with full path for saving the list of excluded colonies must be specified.}
}
\value{
Invisible numeric vector of excluded colonies from neighborless
algorithm.
}
\description{
Neighborless algorithm detects the pinning error of a plate by first excluding
the empty indices and then checks for colony area. If the colony area is
zero, then the algorithm concludes it as one of excluded colonies.
}
\examples{
plateformat <- 1536
data_area <- simulated_data_1536(data_384 = colonyarea$data_subtypes,
                                 in_data_flow = "across",
                                 out_data_flow = "down",
                                 is_plate_coords = TRUE)
empty_indices <- which(convert_small_to_large(plate_from = 384,
                                              plate_to = 1536,
                                              data_from = colonyarea$data_subtypes,
                                              in_data_flow = 'across',
                                              out_data_flow = "down",
                                              is_plate_coords = FALSE)$y \%in\% 'Empty')
(neighborless_algorithm(plateformat          = 1536,
                        colony_area_raw_data = data_area$y,
                        empty_indices        = empty_indices,
                        excluded_colonies    = c(),
                        is_save              = FALSE,
                        excluded_file        = NULL))

}

