% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_middle_indices.R
\name{not_middle_indices}
\alias{not_middle_indices}
\title{Colony indices at the periphery of a given plateformat}
\usage{
not_middle_indices(plateformat)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536}
}
\value{
numeric vector of colony indices at the periphery of the
\code{plateformat}
}
\description{
Colony indices at the periphery of a given plateformat
}
\examples{
not_middle_indices(1536)
}

