% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_middle_neighbors.R
\name{not_middle_neighbors}
\alias{not_middle_neighbors}
\title{Neighbors adjoining the selected colony from the periphery of the given
plateformat}
\usage{
not_middle_neighbors(plateformat, colony)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536}

\item{colony}{An integer which can be one from
\code{not_middle_indices(plateformat)}}
}
\value{
A numeric vector with neighbors adjoining the given \code{colony}
}
\description{
Neighbors adjoining the selected colony from the periphery of the given
plateformat
}
\examples{
not_middle_neighbors(plateformat, 1)
}

