% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinerrordetector.R
\docType{package}
\name{pinerrordetector}
\alias{pinerrordetector}
\alias{pinerrordetector-package}
\title{pinerrordetector}
\description{
Neighbor based algorithms used to detect robotic pinning errors in
yeast genetic interaction screening assay
}
\details{
This package contains functions for detecting pinning errors and
displaying them as platemaps. Both neighborful and neighborless algorithms
are implemented in this R package.

For a complete list of functions,
use \code{library(help = "pinerrordetector")}.
}
\author{
Sathish Kumar Srinivasan

Maintainer: Sathish Kumar Srinivasan \email{sathishsrinivasank@gmail.com}
}

