% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_coords.R
\name{plate_coords}
\alias{plate_coords}
\title{Generate Rectangular Plate Coordinates}
\usage{
plate_coords(plate_to, data_from, data_format)
}
\arguments{
\item{plate_to}{An integer giving plateformat}

\item{data_from}{vector of data for which coordinates are created}

\item{data_format}{A string indicating the plate format of
\code{data_from}. It can take a value either down or across}
}
\value{
Data frame with plate coordinates
}
\description{
plate_coords generates rectangular plate coordinates
}
\examples{
plate_coords(plate_to = 1536,
             data_from = 1:1536,
             data_format = "across")

}

