% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_letters.R
\name{plate_letters}
\alias{plate_letters}
\title{Get Plate letters for row coordinates of a plate}
\usage{
plate_letters()
}
\value{
character vector of row coordinates for a plate
}
\description{
Get Plate letters for row coordinates of a plate
}
\examples{
plate_letters()[1:plate_nrow(1536)]
plate_letters()[1:plate_nrow(6144)]
}

