% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_ncol.R
\name{plate_ncol}
\alias{plate_ncol}
\title{Number of columns for a given plateformat}
\usage{
plate_ncol(plateformat)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536}
}
\value{
An integer with the number of columns for a given \code{plateformat}
}
\description{
Number of columns for a given plateformat
}
\examples{
plate_ncol(1536)
plate_ncol(384)
}

