% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_nrow.R
\name{plate_nrow}
\alias{plate_nrow}
\title{Number of rows for a given plateformat}
\usage{
plate_nrow(plateformat)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536}
}
\value{
An integer with the number of rows for a given \code{plateformat}
}
\description{
Number of rows for a given plateformat
}
\examples{
plate_nrow(1536)
plate_nrow(96)
}

