% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_line_indices.R
\name{right_line_indices}
\alias{right_line_indices}
\title{Get the right line indices of a given plateformat}
\usage{
right_line_indices(plateformat)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536}
}
\value{
The numeric vector with right line indices of  a given
\code{plateformat}
}
\description{
Get the right line indices of a given plateformat
}
\examples{
right_line_indices(384)
right_line_indices(1536)
}

