% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_line_neighbors.R
\name{right_line_neighbors}
\alias{right_line_neighbors}
\title{Get indices of five neighbors for an index at the right line of the plate}
\usage{
right_line_neighbors(plateformat, n)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536}

\item{n}{An integer which can be one of \code{right_line_indices(plateformat)}}
}
\value{
numeric vector of five neighbors for the given \code{plateformat} and
\code{n}
}
\description{
Get indices of five neighbors for an index at the right line of the plate
}
\examples{
right_line_neighbors(384, 370)
right_line_neighbors(1536, 1506)
}

