% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_line_indices.R
\name{top_line_indices}
\alias{top_line_indices}
\title{Get top line indices of a given plateformat}
\usage{
top_line_indices(plateformat)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536}
}
\value{
The numeric vector with top line indices of  a given
\code{plateformat}
}
\description{
Get top line indices of a given plateformat
}
\examples{
top_line_indices(384)
top_line_indices(1536)
}

