% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_line_neighbors.R
\name{top_line_neighbors}
\alias{top_line_neighbors}
\title{Get indices of five neighbors for an index at the top line of the plate}
\usage{
top_line_neighbors(plateformat, n)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536}

\item{n}{An integer which can be one of \code{top_line_indices(plateformat)}}
}
\value{
numeric vector of five neighbors for the given \code{plateformat} and
\code{n}
}
\description{
Get indices of five neighbors for an index at the top line of the plate
}
\examples{
top_line_neighbors(384, 17)
top_line_neighbors(1536, 33)
}

