% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topleft_corner_neighbors.R
\name{topleft_corner_neighbors}
\alias{topleft_corner_neighbors}
\title{Get indices of three neighbors for the top left corner index of a plate}
\usage{
topleft_corner_neighbors(plateformat)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536}
}
\value{
numeric vector of three neighbors for a given \code{plateformat}
}
\description{
Get indices of three neighbors for the top left corner index of a plate
}
\examples{
topleft_corner_neighbors(96)
topleft_corner_neighbors(1536)
}

