% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topright_corner_neighbors.R
\name{topright_corner_neighbors}
\alias{topright_corner_neighbors}
\title{Get indices of three neighbors for the top right corner index of a plate}
\usage{
topright_corner_neighbors(plateformat)
}
\arguments{
\item{plateformat}{An integer which can be one of 96 or 384 or 1536}
}
\value{
numeric vector of three neighbors for a given \code{plateformat}
}
\description{
Get indices of three neighbors for the top right corner index of a plate
}
\examples{
topright_corner_neighbors(96)
topright_corner_neighbors(1536)
}

