% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation2.R
\name{variation2}
\alias{variation2}
\title{Variation-2}
\usage{
variation2(colony, combin_coords, neighbors_selected_colony, excluded_colonies,
  colony_area_raw_data, plate_median_threshold, param1_threshold,
  param2_threshold)
}
\arguments{
\item{colony}{selected colony by the outer loop of the neighborful algorithm}

\item{combin_coords}{combinations of neighbors adjoining the selected colony}

\item{neighbors_selected_colony}{neighbors adjoining the selected colony}

\item{excluded_colonies}{A numeric vector of indices that are part of excluded
colonies}

\item{colony_area_raw_data}{A numeric vector of raw data representing the area
yeast grown at a specific location on a nutrient medium agar plate}

\item{plate_median_threshold}{This threshold value is computed by multiplying
the value of \code{percent_median_thresh} with the value from
\code{plate_median(empty_indices,colony_area_raw_data)} and dividing it by 100}

\item{param1_threshold}{The numeric value which can be one from \code{1:8}.
This exclusion criteria-1 of neighborful algorithm represents the number of
adjoining neighbors of the selected colony from \code{colony_indices} that
has colony area less than or equal to \code{plate_median_threshold}. If the
selected colony is surrounded by at least \code{param1_threshold} neighbors,
then it will be considered as excluded colony by the neighborful algorithm}

\item{param2_threshold}{The numeric value which can be one from \code{1:8}.
This exclusion criteria-2 of neighborful algorithm represents the number of
adjoing neighbors of the selected colony from \code{colony_indices} are
excluded by the previous iteration of neighborful algorithm. If the selected
colony is surrounded by at least \code{param2_threshold} neighbors, then it
will be considered as excluded colony by the neighborful algorithm.}
}
\value{
A vector of excluded colonies by variation2
}
\description{
Inner loop of neighborful algorithm used by its variations B and F
}
\examples{
plateformat <- 1536
middle_colony <- 34
param1_threshold <- 1
param2_threshold <- 1
neighbors_selected_colony <- middle_neighbors(plateformat, middle_colony)
combin_coords <- combn(length(neighbors_selected_colony), param1_threshold)
data_area <- simulated_data_1536(colonyarea$data_subtypes)
empty_indices <- which(convert_384_to_1536_data(colonyarea$data_subtypes)
                       \%in\% 'Empty')
p_median <- plate_median(empty_indices, data_area)
plate_median_threshold <- p_median/4
variation2(colony = middle_colony,
           combin_coords = combin_coords,
           neighbors_selected_colony = neighbors_selected_colony,
           excluded_colonies = c(2),
           colony_area_raw_data = data_area,
           plate_median_threshold = plate_median_threshold,
           param1_threshold = param1_threshold,
           param2_threshold = param2_threshold)
}

