% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend2.R
\name{legend2}
\alias{legend2}
\title{legend2}
\usage{
legend2(x, y = NULL, legend, fill = NULL, col = par("col"),
  border = "black", lty, lwd, pch, angle = 45, density = NULL,
  bty = "o", bg = par("bg"), box.lwd = par("lwd"), box.lty = par("lty"),
  box.col = par("fg"), pt.bg = NA, cex = 1, pt.cex = cex,
  pt.lwd = lwd, xjust = 0, yjust = 1, x.intersp = 1, y.intersp = 1,
  adj = c(0, 0.5), text.width = NULL, text.col = par("col"),
  text.font = NULL, merge = do.lines && has.pch, trace = FALSE,
  plot = TRUE, ncol = 1, horiz = FALSE, title = NULL, inset = 0, xpd,
  title.col = text.col, title.adj = 0.5, seg.len = 2)
}
\value{
returns legend symbol and text
}
\description{
This is a modified function of legend found in graphics package. Lines 84-86
are commented out to prevent xycoords from determining the legend coordinates.
All parameters are the same as given in the legend function of graphics
package of version 3.1.2.
}

