% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_median.R
\name{plate_median}
\alias{plate_median}
\title{Plate median of the given colony area of a plate}
\usage{
plate_median(empty_indices = NULL, colony_area_raw_data, na.rm = TRUE)
}
\arguments{
\item{empty_indices}{A numeric vector of indexed empty locations. It can also
be the indices of control colonies grown on a plate, which may be used for
normalizing the colony area of a plate}

\item{colony_area_raw_data}{A numeric vector of raw data representing the area
yeast grown at a specific location on a nutrient medium agar plate}
}
\value{
Numeric value with plate_median of the given
\code{colony_area_raw_data}
}
\description{
Plate median of the given colony area of a plate. It removes colony area of
empty indices before computing the median of the plate.
}
\examples{
data_area <- simulated_data_1536(data_384 = colonyarea$data_subtypes,
                                 in_data_flow = "across",
                                 out_data_flow = "down",
                                 is_plate_coords = TRUE)
empty_indices <- which(convert_small_to_large(plate_from = 384,
                                              plate_to = 1536,
                                              data_from = colonyarea$data_subtypes,
                                              in_data_flow = 'across',
                                              out_data_flow = "down",
                                              is_plate_coords = FALSE)$y \%in\% 'Empty')
plate_median(empty_indices, data_area$y)   # 600

}

