% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data_1536.R
\name{simulated_data_1536}
\alias{simulated_data_1536}
\title{Simulated data based on data subtypes from 384 to 1536 plateformat}
\usage{
simulated_data_1536(data_subtypes_384, out_data_flow = "across",
  is_plate_coords = TRUE)
}
\arguments{
\item{data_subtypes_384}{A character vector with data subtypes of length 384}

\item{out_data_flow}{A string which tells the function to output the
converted data to a plate configuration which can be down or across}

\item{is_plate_coords}{logical returns plate coordinates, if its value is
\code{TRUE}}
}
\value{
A numeric vector with uniformly distributed data having median of 600
}
\description{
Uniformly distributed data with median of 600 is simulated and returned
for  1536 plateformat configuration from 384 plateformat configuration.
}
\examples{
simulated_data_1536(colonyarea$data_subtypes)
}

